requirements.txt
# requirements.txt

Django>=4.2.15,<4.3
djangorestframework>=3.15.2,<3.16
psycopg2-binary>=2.8.6
drf-spectacular>=0.27.2,<0.28